requirements.txt
fastapi
uvicorn
torch==2.5.1
torchvision
numpy
pillow
matplotlib
pytorch-ignite
scikit-learn
python-multipart
requests
supabase